Added Test cases for new functions
